requirements.txt
olefile
pdfrw
msoffcrypto-tool
